requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
sounddevice==0.4.7
paho-mqtt==1.6.1
